% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\docType{data}
\name{A}
\alias{A}
\title{Simulating}
\format{An object of class \code{matrix} with 50 rows and 50 columns.}
\usage{
A
}
\description{
This script allows you to simulate network data and perform inference via the Expectation Maximization algorithm
}
\keyword{datasets}
