% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_comparison.R
\name{analyse_results}
\alias{analyse_results}
\title{Network comparison}
\usage{
analyse_results(theta = 0.5)
}
\description{
This script plots the algorithm's guess of the underlying network and compares it to the original network A
}
