% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{as.data.frame.network}
\alias{as.data.frame.network}
\title{Coerce a Network Object to a \code{data.frame}}
\usage{
\method{as.data.frame}{network}(
  x,
  ...,
  unit = c("edges", "vertices"),
  na.rm = TRUE,
  attrs_to_ignore = "na"
)
}
\arguments{
\item{x}{an object of class \code{network}}

\item{...}{additional arguments}

\item{unit}{whether a \code{data.frame} of edge or vertex attributes
should be returned.}

\item{na.rm}{logical; ignore missing entries when constructing the data frame?}

\item{attrs_to_ignore}{character; a vector of attribute names to exclude from 
the returned \code{data.frame} (Default: \code{"na"})}
}
\description{
The \code{as.data.frame} method coerces its input to a \code{data.frame} containing
\code{x}'s edges or vertices.
}
