% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_variable_exchange_test.R
\name{benchmark}
\alias{benchmark}
\title{evaluating the exact-approximate MCMC algorithm on a concrete example for which all computations are easy
we want to estimate the precision parameter of a zero-mean univariate gaussian, based on one observation}
\usage{
benchmark(yobs, initial_theta, mcmc_iter, step = 0.5)
}
\arguments{
\item{yobs}{one observation from a normal distribution}

\item{initial_theta}{The initial guess on the value of the precision of the normal}

\item{mcmc_iter}{The number of iterations of the exact-approximate MCMC algorithm}

\item{step}{The standard deviation used in the random walk}
}
\description{
evaluating the exact-approximate MCMC algorithm on a concrete example for which all computations are easy
we want to estimate the precision parameter of a zero-mean univariate gaussian, based on one observation
}
