% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_metropolis_hastings.R
\name{DMH}
\alias{DMH}
\title{This function makes it possible to fit an ERGM model (i.e. sample from the posterior distribution on the parameters) using an enhanced MCMC technique: the Double Metropolis Hastings Algorithm.}
\usage{
DMH(Y, COV, thetas, mcmc_iter, gibbs_cycles, a)
}
\arguments{
\item{Y}{one observation arising from the unknown network structure (must be an adjacency matrix)}

\item{COV}{scaling of the random walk.}

\item{thetas}{a matrix storing the updates of the parameters for each DMH iteration. The first line of the matrix contains the initial values of the parameters.}

\item{mcmc_iter}{The number of iterations of the outer Metropolis Hastings sampler}

\item{gibbs_cycles}{The number of iterations of the inner sampler (the number of Gibbs updates before considering the last state as a draw from the likelihood P(x|theta) for the auxiliary variable sample)}

\item{a}{the scale of the uniform prior for each parameter}
}
\description{
This function makes it possible to fit an ERGM model (i.e. sample from the posterior distribution on the parameters) using an enhanced MCMC technique: the Double Metropolis Hastings Algorithm.
}
