% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generator.R
\name{ERGM}
\alias{ERGM}
\title{This function generates a graph from an ERGM}
\usage{
ERGM(n = 50, p = 0.1)
}
\arguments{
\item{n}{The number of nodes in the network}

\item{p}{The probability of having an edge at any position (i,j)}
}
\description{
This function generates a graph from an ERGM
}
