% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_comparison.R
\name{F_measure_plot}
\alias{F_measure_plot}
\title{Network comparison}
\usage{
F_measure_plot(k, n, p, apha = 0.6, beta = 0.009, rho = 0.1)
}
\description{
This script plots the algorithm's guess of the underlying network and compares it to the original network A
}
