% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-package.R
\name{flo}
\alias{flo}
\title{Florentine Wedding Data (Padgett)}
\source{
Padgett, John F.  (1994). \dQuote{Marriage and Elite Structure in
Renaissance Florence, 1282-1500.} Paper delivered to the Social Science
History Association.
}
\usage{
data(flo)
}
\description{
This is a data set of Padgett (1994), consisting of weddings among leading
Florentine families.  This data is stored in symmetric adjacency matrix
form.
}
\examples{

data(flo)
nflo<-network(flo,directed=FALSE)    #Convert to network object form
all(nflo[,]==flo)                    #Trust, but verify
                                     #A fancy display:
plot(nflo,displaylabels=TRUE,boxed.labels=FALSE,label.cex=0.75)

}
\references{
Wasserman, S. and Faust, K. (1994) \emph{Social Network
Analysis: Methods and Applications}, Cambridge: Cambridge University Press.
}
\seealso{
\code{\link{network}}
}
\keyword{datasets}
