% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_metropolis_hastings.R
\name{Gibbs}
\alias{Gibbs}
\title{This function performs m full Gibbs update of an initial graph Y}
\usage{
Gibbs(Y, theta, m)
}
\arguments{
\item{Y}{The initial graph which is updated m times with Gibbs, to get a draw from the likelihood}

\item{theta}{The parameters of the likelihood.}

\item{m}{The number of full pass over Y's components, i.e. the number of Gibbs updates}
}
\description{
This function performs m full Gibbs update of an initial graph Y
}
