% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generator.R
\name{interact}
\alias{interact}
\title{This function generates noisy data}
\usage{
interact(A, alpha = 0.6, beta = 0.009, n = 50, k)
}
\arguments{
\item{A}{The underlying ERGM}

\item{alpha}{The true positive rate, i.e. the probability of observing an edge at (i,j) when A_{(i,j)}=1}

\item{beta}{The false positive rate, i.e. the probability of observing an edge at (i,j) when A_{(i,j)}=0}

\item{n}{The number of nodes in the network}

\item{k}{The number of observations for each pair (i,j)}
}
\description{
This function generates noisy data
}
