% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mixingmatrix}
\alias{mixingmatrix}
\alias{mixingmatrix.network}
\alias{print.mixingmatrix}
\title{Mixing matrix}
\usage{
mixingmatrix(object, ...)

\method{mixingmatrix}{network}(object, attrname, ...)

\method{print}{mixingmatrix}(x, ...)
}
\arguments{
\item{object}{a network or some other data structure for which a mixing
matrix is meaningful.}

\item{...}{further arguments passed to or used by methods.}

\item{attrname}{a vertex attribute name.}

\item{x}{mixingmatrix object}
}
\description{
Return the mixing matrix for a network, on a given attribute.
}
\examples{
# Interaction ties between Lake Pomona SAR organizations by sponsorship type
# of tie sender and receiver (data from Drabek et al. 1981)
data(emon)
mixingmatrix(emon$LakePomona, "Sponsorship")
}
