\name{network-internal}
\title{Internal Network Package Functions}

%\alias{DistMatrix}
\alias{as.color}
\alias{is.color}
\alias{is.discrete}
\alias{is.discrete.character}
\alias{is.discrete.numeric}
\alias{+}
\alias{-}
\alias{*}
\alias{+.default}
\alias{-.default}
\alias{*.default}
\alias{summary.character}
\alias{print.summary.character}

\description{
  Internal network functions.
}
\usage{
as.color(x)
is.color(x)
is.discrete(x)
is.discrete.character(x)
is.discrete.numeric(x)
}
\details{
  Most of these are not to be called by the user.
}
\arguments{
 \item{x}{an object, to be designated either discrete or continuous.}
%\item{\dots}{further arguments passed to or used by methods.}
}
\seealso{network}
\keyword{internal}
