% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.R
\name{network_from_data_frame}
\alias{network_from_data_frame}
\title{Coercion from Data Frames to Network Objects}
\usage{
network_from_data_frame(edges, directed = TRUE, vertices = NULL,
  loops = FALSE, multiple = FALSE, ...)
}
\arguments{
\item{edges}{A data frame containing the from/to edge list in the first two columns
(the values of which correspond to \code{"vertex.names"}). Additional columns are 
assigned to edge attributes named after those columns.}

\item{directed}{logical, default: \code{TRUE}; should edges be interpreted as directed?}

\item{vertices}{An optional data frame containing the vertex attributes. The first
column is assigned to the \code{"vertex.names"} with additional columns being assigned 
to attributes named after those columns.}

\item{loops}{logical, default: \code{FALSE}; should loops be allowed?}

\item{multiple}{logical, default: \code{FALSE}; are multiplex edges allowed?}

\item{...}{Arguments passed to or from other methods.}
}
\value{
An object of class \code{network}
}
\description{
Coercion from Data Frames to Network Objects
}
\examples{
vertex_df <- data.frame(name = letters[1:5],
                        int_attr = seq_len(5),
                        chr_attr = LETTERS[1:5],
                        lgl_attr = c(TRUE, FALSE, TRUE, FALSE, TRUE),
                        stringsAsFactors = FALSE)
vertex_df                        


edge_df <- data.frame(from = c("b", "c", "c", "d", "d", "e"),
                      to = c("a", "b", "a", "a", "b", "a"),
                      int_attr = seq_len(6),
                      chr_attr = LETTERS[1:6],
                      lgl_attr = c(TRUE, FALSE, TRUE, FALSE, TRUE, FALSE),
                      stringsAsFactors = FALSE)
edge_df

g <- network_from_data_frame(edge_df, directed = TRUE, vertices = vertex_df)
g

}
\author{
Brendan Knapp \email{brendan.knapp@nps.edu}
}
