% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generator.R
\name{sampleErdosRenyi}
\alias{sampleErdosRenyi}
\title{This function generates an Erdos Renyi graph, the model used in the paper}
\usage{
sampleErdosRenyi(n = 50, p = 0.1)
}
\arguments{
\item{n}{The number of nodes in the network}

\item{p}{The probability of having an edge at any position (i,j)}
}
\description{
This function generates an Erdos Renyi graph, the model used in the paper
}
\examples{
sampleErdosRenyi_function()
}
