% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_infer}
\alias{simulate_infer}
\title{Simulating}
\usage{
simulate_infer(n = 50, k = 10, p = 0.1, alpha = 0.6,
  beta = 0.009)
}
\description{
This function allows you to simulate network data and perform inference via the Expectation Maximization algorithm
}
