% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_metropolis_hastings.R
\name{Summary}
\alias{Summary}
\title{This function computes the statistics of the graph A}
\usage{
Summary(A)
}
\arguments{
\item{A}{The adjacency matrix of the graph of which we want to compute the statistics}
}
\description{
This function computes the statistics of the graph A
}
