% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_metropolis_hastings.R
\name{trunc_norm}
\alias{trunc_norm}
\title{This function makes it possible to fit an ergm model (i.e. sample from the posterior distribution on the parameters) using an enhanced MCMC technique.}
\usage{
trunc_norm(old_theta, lower_theta, upper_theta)
}
\arguments{
\item{lower_theta}{Used to bound the possible values of the parameters}

\item{upper_theta}{Used to bound the possible values of the parameters}

\item{yobs}{one observation of the network}

\item{initial_theta}{The initial guess on the values of the parameters of the ergm model}

\item{mcmc_iter}{The number of iterations of the MCMC algorithm}
}
\description{
This function makes it possible to fit an ergm model (i.e. sample from the posterior distribution on the parameters) using an enhanced MCMC technique.
}
