% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generator.R
\name{ERGM}
\alias{ERGM}
\title{This function generates a graph from an ERGM}
\usage{
ERGM(n, param)
}
\arguments{
\item{n}{The number of nodes in the network}
}
\description{
This function generates a graph from an ERGM
}
