% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R, R/operators.R
\name{network-internal}
\alias{network-internal}
\alias{is.discrete.numeric}
\alias{is.discrete.character}
\alias{is.discrete}
\alias{networkOperatorSetup}
\title{Internal Network Package Functions}
\usage{
is.discrete.numeric(x)

is.discrete.character(x)

is.discrete(x)

networkOperatorSetup(x, y = NULL)
}
\arguments{
\item{x}{an object to be designated either discrete or continuous, or a
network.}

\item{y}{a network or something coercible to one.}
}
\description{
Internal network functions.
}
\details{
Most of these are not to be called by the user.
}
\seealso{
network
}
\keyword{internal}
