% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.R
\name{network_from_data_frame}
\alias{network_from_data_frame}
\title{Coercion from Data Frames to Network Objects}
\usage{
network_from_data_frame(edges, directed = TRUE, vertices = NULL,
  loops = FALSE, multiple = FALSE, ...)
}
\arguments{
\item{edges}{data frame containing the from/to edge list in the first two columns, with
additional columns treated as edge attributes.}

\item{directed}{logical, default: `TRUE`; should edges be interpreted as directed?}

\item{vertices}{optional data frame containing the vertex attributes. The first column 
is assigned to the `"vertex.names"` attribute with additional columns using column
 names.}

\item{loops}{logical, default: `FALSE`; should loops be allowed?}

\item{multiple}{logical; are multiplex edges allowed?}

\item{...}{additional arguments}
}
\value{
An object of class \code{network}
}
\description{
Coercion from Data Frames to Network Objects
}
\examples{
vertex_df <- data.frame(name = letters[1:5],
                        int_attr = seq_len(5),
                        chr_attr = LETTERS[1:5],
                        lgl_attr = c(TRUE, FALSE, TRUE, FALSE, TRUE),
                        stringsAsFactors = FALSE)
vertex_df[["df_list_attr"]] <- replicate(5, mtcars, simplify = FALSE)

edge_df <- data.frame(from = c("b", "c", "c", "d", "d", "e"),
                      to = c("a", "b", "a", "a", "b", "a"),
                      int_attr = seq_len(6),
                      chr_attr = LETTERS[1:6],
                      lgl_attr = c(TRUE, FALSE, TRUE, FALSE, TRUE, FALSE),
                      stringsAsFactors = FALSE)
edge_df[["df_list_attr"]] <- replicate(6, mtcars, simplify = FALSE)

g <- network_from_data_frame(edge_df, directed = TRUE, vertices = vertex_df)

}
\seealso{
\code{\link{edgeset.constructors}}, \code{\link{network}},
}
\author{
Brendan Knapp \email{brendan.knapp@nps.edu}
}
\keyword{classes}
\keyword{graphs}
